split up handlers and simulate files
