+++ docs/api_reference/developer/handlers.md
@@ -6,7 +6,13 @@
       show_root_toc_entry: false
       members:
         - condition
-        - log_prob
         - sample
-        - simulate
         - plate
+
+::: dynestyx.api
+    options:
+      show_root_heading: false
+      show_root_toc_entry: false
+      members:
+        - log_prob
+        - simulate
