+++ dynestyx/__init__.py
@@ -4,8 +4,9 @@ from importlib.metadata import version
 
 __version__ = version("dynestyx")
 
+from dynestyx.api import log_prob, simulate
 from dynestyx.discretizers import Discretizer, euler_maruyama
-from dynestyx.handlers import condition, log_prob, plate, sample, simulate
+from dynestyx.handlers import condition, plate, sample
 from dynestyx.inference.configs.simulator import (
     ODESimulatorConfig,
     SDESimulatorConfig,
