+++ dynestyx/handlers.py
@@ -1,26 +1,20 @@
-"""Contains the `sample` and `infer` primitives and `effectful` utilities for `dynestyx`."""
+"""Contains the core dynestyx primitives and `effectful` handler utilities."""
 
 from typing import Any, TypeVar
 
-import jax.numpy as jnp
 import numpyro
 from effectful.ops.semantics import fwd, handler
 from effectful.ops.syntax import ObjectInterpretation, defop, implements
 from effectful.ops.types import NotHandled
 from jaxtyping import Array, Bool, Real
 
-from dynestyx.inference.latent.trajectory_log_probs import (
-    compute_trajectory_log_prob_terms,
-)
 from dynestyx.models import (
     DynamicalModel,
 )
 from dynestyx.observation_missingness import (
-    MissingObservationMetadata,
-    MissingObservationStrategy,
     prepare_observation_views,
 )
-from dynestyx.types import FunctionOfTime, SimulatedResult
+from dynestyx.types import FunctionOfTime
 from dynestyx.utils import (
     _get_dynamics_with_t0,
     _validate_control_dim,
@@ -113,192 +107,6 @@ def _validate_and_prepare(
     return dynamics_with_t0, obs_values_filled, obs_mask, obs_has_missing
 
 
-def condition(
-    name: str,
-    dynamics: DynamicalModel,
-    *,
-    obs_times: Real[Array, "*obs_time_plate obs_time"] | None = None,
-    obs_values: Real[Array, "*obs_value_plate obs_time observation_dim"]
-    | Real[Array, "*obs_value_plate obs_time"]
-    | None = None,
-    ctrl_times: Real[Array, "*ctrl_time_plate ctrl_time"] | None = None,
-    ctrl_values: Real[Array, "*ctrl_value_plate ctrl_time control_dim"]
-    | Real[Array, "*ctrl_value_plate ctrl_time"]
-    | None = None,
-    predict_times: Real[Array, "*predict_time_plate predict_time"] | None = None,
-    **kwargs,
-):
-    """Run inference on a dynamical model without registering numpyro sites.
-
-    This is the numpyro-free entry point. When a Filter or Smoother handler
-    is active, returns a ConditionedResult dataclass with marginal_loglik, states, etc.
-
-    Parameters:
-        name: Name of the inference site.
-        dynamics: Dynamical model to infer.
-        obs_times: Times at which observations are available.
-        obs_values: Values of the observations at the given times.
-        ctrl_times: Times at which controls are applied.
-        ctrl_values: Values of the controls at the given times.
-        predict_times: Times at which to predict.
-        **kwargs: Additional keyword arguments.
-    """
-    dynamics_with_t0, obs_values_filled, obs_mask, obs_has_missing = (
-        _validate_and_prepare(
-            name,
-            dynamics,
-            obs_times=obs_times,
-            obs_values=obs_values,
-            ctrl_times=ctrl_times,
-            ctrl_values=ctrl_values,
-            predict_times=predict_times,
-        )
-    )
-
-    return _condition_intp(
-        name,
-        dynamics_with_t0,
-        obs_times=obs_times,
-        obs_values=obs_values,
-        _obs_values_filled=obs_values_filled,
-        _obs_mask=obs_mask,
-        _obs_has_missing=obs_has_missing,
-        ctrl_times=ctrl_times,
-        ctrl_values=ctrl_values,
-        predict_times=predict_times,
-        **kwargs,
-    )
-
-
-def log_prob(
-    dynamics: DynamicalModel,
-    *,
-    state_path_params,
-    state_path_param_times,
-    obs_times: Real[Array, "*obs_time_plate obs_time"] | None = None,
-    obs_values: Real[Array, "*obs_value_plate obs_time observation_dim"]
-    | Real[Array, "*obs_value_plate obs_time"]
-    | None = None,
-    ctrl_times: Real[Array, "*ctrl_time_plate ctrl_time"] | None = None,
-    ctrl_values: Real[Array, "*ctrl_value_plate ctrl_time control_dim"]
-    | Real[Array, "*ctrl_value_plate ctrl_time"]
-    | None = None,
-    missing_observation_strategy: MissingObservationStrategy = "auto",
-    missing_obs_values=None,
-    missing_obs_metadata: MissingObservationMetadata | None = None,
-    chunk_size: int | None = None,
-    ode_diffeqsolve_settings=None,
-):
-    """Return the pure-JAX joint log density for a state path and observations.
-
-    Parameters:
-        dynamics: Dynamical model to score.
-        state_path_params: Concrete path-parameter values supplied by the
-            caller. Writing ``z = state_path_params`` and
-            ``x = g(z) = state_path``, this function evaluates ``log p(x, y)``
-            after reconstructing the path ``x``. For discrete / discretized
-            models ``z`` is the full latent path in v1. For ODE models ``z``
-            is the initial condition in v1.
-        state_path_param_times: Times attached to ``state_path_params``.
-        obs_times: Times at which observations are available.
-        obs_values: Observation values at ``obs_times``.
-        ctrl_times: Times at which controls are supplied.
-        ctrl_values: Control values aligned to ``ctrl_times``.
-        missing_observation_strategy: Strategy for handling missing
-            observation coordinates. `"auto"` prefers exact marginalization when
-            supported and otherwise falls back to explicit augmentation for
-            continuous observation families.
-        missing_obs_values: Explicit values for the missing observation
-            coordinates when augmentation is active. In that case this function
-            scores the augmented complete-data target
-            ``log p(x, y_observed, y_missing | ...)`` rather than the
-            marginalized observed-data target.
-        missing_obs_metadata: Optional precomputed metadata defining the flat
-            ordering of ``missing_obs_values`` for traced/JIT callers.
-        chunk_size: Optional host-level chunk size for scoring loops.
-        ode_diffeqsolve_settings: Optional Diffrax solve settings used when
-            reconstructing deterministic continuous-time trajectories.
-    """
-    state_path_param_times = jnp.asarray(state_path_param_times)
-    dynamics_with_t0, obs_values_filled, obs_mask, _obs_has_missing = (
-        _validate_and_prepare(
-            "log_prob",
-            dynamics,
-            obs_times=obs_times,
-            obs_values=obs_values,
-            ctrl_times=ctrl_times,
-            ctrl_values=ctrl_values,
-            predict_times=state_path_param_times,
-        )
-    )
-
-    terms = compute_trajectory_log_prob_terms(
-        dynamics_with_t0,
-        state_path_params=state_path_params,
-        state_path_param_times=state_path_param_times,
-        obs_times=obs_times,
-        obs_values=obs_values,
-        obs_values_filled=obs_values_filled,
-        obs_mask=obs_mask,
-        missing_observation_strategy=missing_observation_strategy,
-        missing_obs_values=missing_obs_values,
-        missing_obs_metadata=missing_obs_metadata,
-        ctrl_times=ctrl_times,
-        ctrl_values=ctrl_values,
-        chunk_size=chunk_size,
-        ode_diffeqsolve_settings=ode_diffeqsolve_settings,
-    )
-    return terms.joint_log_prob
-
-
-def simulate(
-    dynamics: DynamicalModel,
-    *,
-    rng_key,
-    obs_times: Real[Array, "*obs_time_plate obs_time"] | None = None,
-    ctrl_times: Real[Array, "*ctrl_time_plate ctrl_time"] | None = None,
-    ctrl_values: Real[Array, "*ctrl_value_plate ctrl_time control_dim"]
-    | Real[Array, "*ctrl_value_plate ctrl_time"]
-    | None = None,
-    predict_times: Real[Array, "*predict_time_plate predict_time"] | None = None,
-    n_simulations: int = 1,
-    simulator_config=None,
-    **simulator_kwargs,
-) -> SimulatedResult:
-    """Run pure-JAX forward simulation and return a :class:`SimulatedResult`.
-
-    Unlike :func:`sample`, this entry point is for data generation only and does
-    not register any NumPyro sample/deterministic sites. The simulator time grid
-    is taken from ``predict_times`` when provided, else from ``obs_times``.
-    """
-    if obs_times is None and predict_times is None:
-        raise ValueError("At least one of obs_times or predict_times must be provided")
-
-    _validate_site_sorting(obs_times, name="obs_times")
-    _validate_site_sorting(ctrl_times, name="ctrl_times")
-    _validate_site_sorting(predict_times, name="predict_times")
-    _validate_controls(obs_times, predict_times, ctrl_times, ctrl_values)
-    _validate_control_dim(dynamics, ctrl_values)
-
-    dynamics_with_t0 = _get_dynamics_with_t0(dynamics, obs_times, predict_times)
-
-    from dynestyx.simulation import Simulator
-
-    simulator = Simulator(
-        n_simulations=n_simulations,
-        simulator_config=simulator_config,
-        **simulator_kwargs,
-    )
-    return simulator.simulate(
-        dynamics_with_t0,
-        rng_key=rng_key,
-        obs_times=obs_times,
-        ctrl_times=ctrl_times,
-        ctrl_values=ctrl_values,
-        predict_times=predict_times,
-    )
-
-
 def sample(
     name: str,
     dynamics: DynamicalModel,
@@ -356,6 +164,63 @@ def sample(
         register_numpyro_sites(name)
 
     return result
+
+
+def condition(
+    name: str,
+    dynamics: DynamicalModel,
+    *,
+    obs_times: Real[Array, "*obs_time_plate obs_time"] | None = None,
+    obs_values: Real[Array, "*obs_value_plate obs_time observation_dim"]
+    | Real[Array, "*obs_value_plate obs_time"]
+    | None = None,
+    ctrl_times: Real[Array, "*ctrl_time_plate ctrl_time"] | None = None,
+    ctrl_values: Real[Array, "*ctrl_value_plate ctrl_time control_dim"]
+    | Real[Array, "*ctrl_value_plate ctrl_time"]
+    | None = None,
+    predict_times: Real[Array, "*predict_time_plate predict_time"] | None = None,
+    **kwargs,
+):
+    """Run inference on a dynamical model without registering numpyro sites.
+
+    This is the numpyro-free entry point. When a Filter or Smoother handler
+    is active, returns a ConditionedResult dataclass with marginal_loglik, states, etc.
+
+    Parameters:
+        name: Name of the inference site.
+        dynamics: Dynamical model to infer.
+        obs_times: Times at which observations are available.
+        obs_values: Values of the observations at the given times.
+        ctrl_times: Times at which controls are applied.
+        ctrl_values: Values of the controls at the given times.
+        predict_times: Times at which to predict.
+        **kwargs: Additional keyword arguments.
+    """
+    dynamics_with_t0, obs_values_filled, obs_mask, obs_has_missing = (
+        _validate_and_prepare(
+            name,
+            dynamics,
+            obs_times=obs_times,
+            obs_values=obs_values,
+            ctrl_times=ctrl_times,
+            ctrl_values=ctrl_values,
+            predict_times=predict_times,
+        )
+    )
+
+    return _condition_intp(
+        name,
+        dynamics_with_t0,
+        obs_times=obs_times,
+        obs_values=obs_values,
+        _obs_values_filled=obs_values_filled,
+        _obs_mask=obs_mask,
+        _obs_has_missing=obs_has_missing,
+        ctrl_times=ctrl_times,
+        ctrl_values=ctrl_values,
+        predict_times=predict_times,
+        **kwargs,
+    )
 
 
 @defop
